+++ Code/diorama_tests_files.cmake
@@ -8,6 +8,7 @@ set(FILES
     Tests/Clients/Camera2DTest.cpp
     Tests/Clients/Particles2DTest.cpp
     Tests/Clients/SpriteCullTest.cpp
+    Tests/Clients/SpriteTrailTest.cpp
     Tests/Clients/AsepriteBinaryTest.cpp
     Tests/Clients/TilemapV2Test.cpp
     Tests/Clients/InputActionMapTest.cpp